requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cuda-python==12.6.0
cycler==0.12.1
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.65.5
h11==0.14.0
idna==3.7
importlib_metadata==8.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.43.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optuna==3.6.1
packaging==24.1
pdfkit==1.0.0
pillow==10.4.0
prometheus_client==0.20.0
protobuf==4.25.4
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-rapidjson==1.20
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.1
torch==2.4.0
torchvision==0.19.0
pandas==1.5.1
scikit-learn==1.5.1